requirements.txt
asgiref==3.7.0
charset-normalizer==3.0.1
Django==5.0.6
django-filter==22.1
djangorestframework==3.14.0
djongo
dnspython==2.3.0
idna==3.4
Markdown==3.4.1
pymongo==3.12.3
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.3.1
tzdata==2022.7
urllib3==1.26.14
wincertstore==0.2
drf_spectacular
django-sslserver   
itsdangerous
python-dotenv
django-storages
Pillow
django-cors-headers 
django-seed
Faker
django-extensions==3.2.1
django-extensions==3.2.1
django-resized
channels
daphne
fuzzywuzzy
django4-background-tasks
djangorestframework_simplejwt
python-Levenshtein
psycopg2-binary
bcrypt
PyJWT
streamlit
extra_streamlit_components
streamlit_option_menu
streamlit_modal
plotly
pandas
openpyxl
numpy
redis
matplotlib
seaborn
openpyxl
category_encoders
impyute
catboost
statsmodels
scikit-learn==1.2.2
imblearn
imbalanced-learn==0.10.1
lightgbm
delayed
streamlit_cookies_controller
streamlit_lottie
django-simple-captcha